extracted copy to clipboard functionality, faqs now have clickable links
