+++ components/ui/Accordion/index.tsx
@@ -1,7 +1,9 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { ReactElement, useEffect, useRef, useState } from 'react';
 
+import { ReactComponent as CopySVG } from '@/assets/svgs/copy.svg';
 import { Document } from '@contentful/rich-text-types';
+import Link from 'next/link';
 import { ReactComponent as MinusSVG } from '@/assets/svgs/minus.svg';
 import { ReactComponent as PlusSVG } from '@/assets/svgs/plus.svg';
 import RichBody from '@/components/RichBody';
@@ -70,6 +72,16 @@ export default function Accordion(props: Props): ReactElement {
           </>
         )}
         {question}
+        <Link href={`#${id}`}>
+          <a className="focus:outline-none">
+            <CopySVG
+              className={classNames(
+                'inline w-4 h-4 fill-current mb-1 mr-2 mt-1 ml-2 opacity-0',
+                'hover:opacity-100'
+              )}
+            />
+          </a>
+        </Link>
       </div>
       <div
         className={classNames(
